refactor: Move chat_created fixture
